Move examples into packages
