+++ app/src/main/java/com/mikescamell/motionlayout/Utils.kt
@@ -1,4 +1,4 @@
-package com.mikescamell.locomotionlayout
+package com.mikescamell.motionlayout
 
 import android.view.View
