+++ app/src/main/java/com/mikescamell/motionlayout/loco_motionlayout/CombinedSceneLol.kt
@@ -1,4 +1,4 @@
-package com.mikescamell.locomotionlayout
+package com.mikescamell.motionlayout.loco_motionlayout
 
 import android.graphics.Color
 import android.graphics.PorterDuff
@@ -8,31 +8,41 @@ import android.view.View
 import android.widget.ImageView
 import androidx.appcompat.app.AppCompatActivity
 import androidx.constraintlayout.motion.widget.MotionLayout
-import androidx.constraintlayout.motion.widget.MotionScene
 import androidx.constraintlayout.motion.widget.TransitionAdapter
 import androidx.core.graphics.ColorUtils
 import androidx.vectordrawable.graphics.drawable.Animatable2Compat
 import androidx.vectordrawable.graphics.drawable.AnimatedVectorDrawableCompat
+import com.mikescamell.motionlayout.R
+import com.mikescamell.motionlayout.calculateProgressAlpha
+import com.mikescamell.motionlayout.setLightStatusBar
 import kotlinx.android.synthetic.main.layout2_part4.*
 
-class CombinedScenePart1 : AppCompatActivity() {
+class CombinedSceneLol : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.layout_combined_part1)
+        setContentView(R.layout.layout_combined_lol)
 
         window.decorView.systemUiVisibility = View.SYSTEM_UI_FLAG_LAYOUT_FULLSCREEN
         setLightStatusBar(window.decorView)
 
         val topLeftAnimationForward =
-            AnimatedVectorDrawableCompat.create(this, R.drawable.top_left_liquid_forward)
+            AnimatedVectorDrawableCompat.create(this,
+                R.drawable.top_left_liquid_forward_lol
+            )
         val topLeftAnimationReverse =
-            AnimatedVectorDrawableCompat.create(this, R.drawable.top_left_liquid_reverse)
+            AnimatedVectorDrawableCompat.create(this,
+                R.drawable.top_left_liquid_reverse_lol
+            )
 
         val bottomRightAnimationForward =
-            AnimatedVectorDrawableCompat.create(this, R.drawable.bottom_right_liquid_forward)
+            AnimatedVectorDrawableCompat.create(this,
+                R.drawable.bottom_right_liquid_forward_lol
+            )
         val bottomRightAnimationReverse =
-            AnimatedVectorDrawableCompat.create(this, R.drawable.bottom_right_liquid_reverse)
+            AnimatedVectorDrawableCompat.create(this,
+                R.drawable.bottom_right_liquid_reverse_lol
+            )
 
         val topLeftImageView = (findViewById<ImageView>(R.id.bookSynopsisFluidTopLeft)).apply {
             setImageDrawable(topLeftAnimationForward)
@@ -51,12 +61,15 @@ class CombinedScenePart1 : AppCompatActivity() {
                 progress: Float
             ) {
                 val color =
-                    ColorUtils.setAlphaComponent(Color.WHITE, calculateProgressAlpha(progress))
+                    ColorUtils.setAlphaComponent(Color.WHITE,
+                        calculateProgressAlpha(progress)
+                    )
                 bottomRightAnimationForward?.setColorFilter(color, PorterDuff.Mode.SRC_ATOP)
                 bottomRightAnimationReverse?.setColorFilter(color, PorterDuff.Mode.SRC_ATOP)
             }
 
         })
+
         topLeftAnimationForward?.registerAnimationCallback(object :
             Animatable2Compat.AnimationCallback() {
             override fun onAnimationEnd(drawable: Drawable?) {
