+++ app/src/main/java/com/mikescamell/motionlayout/loco_motionlayout/CombinedScenePart1.kt
@@ -1,31 +1,48 @@
-package com.mikescamell.locomotionlayout
+package com.mikescamell.motionlayout.loco_motionlayout
 
+import android.graphics.Color
+import android.graphics.PorterDuff
 import android.graphics.drawable.Drawable
 import android.os.Bundle
 import android.view.View
 import android.widget.ImageView
 import androidx.appcompat.app.AppCompatActivity
+import androidx.constraintlayout.motion.widget.MotionLayout
+import androidx.constraintlayout.motion.widget.TransitionAdapter
+import androidx.core.graphics.ColorUtils
 import androidx.vectordrawable.graphics.drawable.Animatable2Compat
 import androidx.vectordrawable.graphics.drawable.AnimatedVectorDrawableCompat
+import com.mikescamell.motionlayout.R
+import com.mikescamell.motionlayout.calculateProgressAlpha
+import com.mikescamell.motionlayout.setLightStatusBar
+import kotlinx.android.synthetic.main.layout2_part4.*
 
-class Scene1Part7 : AppCompatActivity() {
+class CombinedScenePart1 : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.layout1_part7)
+        setContentView(R.layout.layout_combined_part1)
 
         window.decorView.systemUiVisibility = View.SYSTEM_UI_FLAG_LAYOUT_FULLSCREEN
         setLightStatusBar(window.decorView)
 
         val topLeftAnimationForward =
-            AnimatedVectorDrawableCompat.create(this, R.drawable.top_left_liquid_forward)
+            AnimatedVectorDrawableCompat.create(this,
+                R.drawable.top_left_liquid_forward
+            )
         val topLeftAnimationReverse =
-            AnimatedVectorDrawableCompat.create(this, R.drawable.top_left_liquid_reverse)
+            AnimatedVectorDrawableCompat.create(this,
+                R.drawable.top_left_liquid_reverse
+            )
 
         val bottomRightAnimationForward =
-            AnimatedVectorDrawableCompat.create(this, R.drawable.bottom_right_liquid_forward)
+            AnimatedVectorDrawableCompat.create(this,
+                R.drawable.bottom_right_liquid_forward
+            )
         val bottomRightAnimationReverse =
-            AnimatedVectorDrawableCompat.create(this, R.drawable.bottom_right_liquid_reverse)
+            AnimatedVectorDrawableCompat.create(this,
+                R.drawable.bottom_right_liquid_reverse
+            )
 
         val topLeftImageView = (findViewById<ImageView>(R.id.bookSynopsisFluidTopLeft)).apply {
             setImageDrawable(topLeftAnimationForward)
@@ -35,6 +52,23 @@ class Scene1Part7 : AppCompatActivity() {
                 setImageDrawable(bottomRightAnimationForward)
             }
 
+        root.setTransitionListener(object : TransitionAdapter() {
+
+            override fun onTransitionChange(
+                motionLayout: MotionLayout,
+                startId: Int,
+                endId: Int,
+                progress: Float
+            ) {
+                val color =
+                    ColorUtils.setAlphaComponent(Color.WHITE,
+                        calculateProgressAlpha(progress)
+                    )
+                bottomRightAnimationForward?.setColorFilter(color, PorterDuff.Mode.SRC_ATOP)
+                bottomRightAnimationReverse?.setColorFilter(color, PorterDuff.Mode.SRC_ATOP)
+            }
+
+        })
         topLeftAnimationForward?.registerAnimationCallback(object :
             Animatable2Compat.AnimationCallback() {
             override fun onAnimationEnd(drawable: Drawable?) {
