+++ app/src/main/java/com/mikescamell/motionlayout/loco_motionlayout/Scene1Part5.kt
@@ -1,14 +1,16 @@
-package com.mikescamell.locomotionlayout
+package com.mikescamell.motionlayout.loco_motionlayout
 
 import android.os.Bundle
 import android.view.View
 import androidx.appcompat.app.AppCompatActivity
+import com.mikescamell.motionlayout.R
+import com.mikescamell.motionlayout.setLightStatusBar
 
-class Scene1Part1 : AppCompatActivity() {
+class Scene1Part5 : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.layout1_part1)
+        setContentView(R.layout.layout1_part5)
 
         window.decorView.systemUiVisibility = View.SYSTEM_UI_FLAG_LAYOUT_FULLSCREEN
         setLightStatusBar(window.decorView)
