+++ app/src/main/res/layout/activity_loco_motionlayout.xml
@@ -1,0 +1,122 @@
+<?xml version="1.0" encoding="utf-8"?>
+<ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:padding="16dp">
+
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:orientation="vertical">
+
+        <com.google.android.material.button.MaterialButton
+            android:id="@+id/scene1Part1Button"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="Scene 1 - Part 1" />
+
+        <com.google.android.material.button.MaterialButton
+            android:id="@+id/scene1Part2Button"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="Scene 1 - Part 2" />
+
+        <com.google.android.material.button.MaterialButton
+            android:id="@+id/scene1Part3Button"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="Scene 1 - Part 3" />
+
+        <com.google.android.material.button.MaterialButton
+            android:id="@+id/scene1Part4Button"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="Scene 1 - Part 4" />
+
+        <com.google.android.material.button.MaterialButton
+            android:id="@+id/scene1Part5Button"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="Scene 1 - Part 5" />
+
+        <com.google.android.material.button.MaterialButton
+            android:id="@+id/scene1Part6Button"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="Scene 1 - Part 6" />
+
+        <com.google.android.material.button.MaterialButton
+            android:id="@+id/scene1Part7Button"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="Scene 1 - Part 7" />
+
+        <com.google.android.material.button.MaterialButton
+            android:id="@+id/scene2Part1Button"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="Scene 2 - Part 1" />
+
+        <com.google.android.material.button.MaterialButton
+            android:id="@+id/scene2Part2Button"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="Scene 2 - Part 2" />
+
+        <com.google.android.material.button.MaterialButton
+            android:id="@+id/scene2Part3Button"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="Scene 2 - Part 3" />
+
+        <com.google.android.material.button.MaterialButton
+            android:id="@+id/scene2Part4Button"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="Scene 2 - Part 4" />
+
+        <com.google.android.material.button.MaterialButton
+            android:id="@+id/scene2Part5Button"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="Scene 2 - Part 5" />
+
+        <com.google.android.material.button.MaterialButton
+            android:id="@+id/scene2Part6Button"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="Scene 2 - Part 6" />
+
+        <com.google.android.material.button.MaterialButton
+            android:id="@+id/scene2Part7Button"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="Scene 2 - Part 7" />
+
+        <com.google.android.material.button.MaterialButton
+            android:id="@+id/combinedScenePart1"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="Combined Scene - Part 1" />
+
+        <com.google.android.material.button.MaterialButton
+            android:id="@+id/combinedScenePart2"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="Combined Scene - Part 2" />
+
+        <com.google.android.material.button.MaterialButton
+            android:id="@+id/combinedScenePart3"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="Combined Scene - Part 3" />
+
+        <com.google.android.material.button.MaterialButton
+            android:id="@+id/combinedScenePartLol"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="Combined Scene - LOLOLOLOL" />
+
+    </LinearLayout>
+
+</ScrollView>
